Fix `ThreadPoolBulkheadTest` flakiness
All tests expect the first thread to occupy the bulkhead and cause one of the other threads to queue (`queueCapacity = 1`) and the remaining one to be rejected with a `BulkheadFullException`. However, there's a race condition between the three threads. If the second or third thread executes before the first one enters the bulkhead, then all three threads will succeed in performing the operation.

This change adds a check to ensure that the second and third threads are only started after the first thread has entered the bulkhead.